requirement.txt
scipy==1.10.1
librosa==0.9.2
matplotlib==3.7.0
tqdm==4.63
seaborn==0.12.2
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
pyYAML==6.0
timm==0.3.2
